Rename govuk_docker_cli to cli
